+++ qaequilibrae/modules/processing_provider/translatable_algo.py
